src/requirements.txt
# Web Scraping
beautifulsoup4==4.14.3
requests==2.32.5
playwright==1.57.0
aiohttp==3.13.3
Scrapy==2.13.4

# Docs processing
PyPDF2==3.0.1
pytesseract==0.3.13
python-docx==1.2.0
pdf2image==1.17.0
pillow==12.1.0


